Refactor testing types to library.
We should test future<T> with interesting types, e.g., classes that lack
a default constructor. This refactors some types that we already had in
the tests for optional<T> to the testing_util library. Part of the work
for #1345.
