+++ google/cloud/google_cloud_cpp_testing.bzl
@@ -5,6 +5,7 @@ google_cloud_cpp_testing_HDRS = [
     "testing_util/chrono_literals.h",
     "testing_util/environment_variable_restore.h",
     "testing_util/init_google_mock.h",
+    "testing_util/testing_types.h",
 ]
 
 google_cloud_cpp_testing_SRCS = [
@@ -12,4 +13,5 @@ google_cloud_cpp_testing_SRCS = [
     "testing_util/environment_variable_restore.cc",
     "testing_util/custom_google_mock_main.cc",
     "testing_util/init_google_mock.cc",
+    "testing_util/testing_types.cc",
 ]
